+++ includes/class-cocart.php
@@ -27,7 +27,7 @@ final class CoCart {
 	 * @static
 	 * @var string
 	 */
-	public static $version = '3.7.11';
+	public static $version = '3.8.0';
 
 	/**
 	 * CoCart Database Schema version.
@@ -128,7 +128,7 @@ final class CoCart {
 		self::define( 'COCART_REVIEW_URL', 'https://wordpress.org/support/plugin/cart-rest-api-for-woocommerce/reviews/' );
 		self::define( 'COCART_DOCUMENTATION_URL', 'https://docs.cocart.xyz' );
 		self::define( 'COCART_TRANSLATION_URL', 'https://translate.cocart.xyz/projects/cart-rest-api-for-woocommerce/' );
-		self::define( 'COCART_NEXT_VERSION', '3.0.0' );
+		self::define( 'COCART_NEXT_VERSION', '4.0.0' );
 	} // END setup_constants()
 
 	/**